improve illegal lock access logic
still tests still fail for get tests, because dynamic dependants just
randomly access their dependencies. these will be locked once they are
added as actual depenedencies, but that may be after the dynamic
dependency has decided that its update is finished. the question is
whether or not this is a problem (hint probably yes, we might read
somthing before and something else after another transaction).