remove transaction handle dynamic after
thinking about it, all afters are inside reevaluations which just use
ticket based access anyways.
This change does allow hacks that acces dynamic before using the
transaction handle, but that seems to be also an OK usage for FullMV.
Although, that would be a different semantics between fullMV and other
Schedulers.