experiment with different inference/extension paradigm
It seems that IntelliJ can not deal with implicit classes thet have
implicit parameters, which makes the ArdtOpsContains hinder more than it
helps.

Instead, we now use the permissions as the first arguments, which solves
just enough type inference issues that we do not need ArdtOpsContains
anymore to infer the element types.

That only leaves the issue of ambiguous extension method names in case
some object contains multiple RDTs in its parameter list. It seems
reasonable to just explicitly use the syntax object in such cases
as that also always enables the IDE to understand where to search.

It is a bit more explicit in some cases, and all of the nested contexts
still seem to not work well, but otherwise seems to be an improvement