change replica ID strategy to just use given ids instead of permissions
This decreases ergonomics for the test suite a lot, because of how much
harder it is now impossible to manage different replicaIds within the
containing datastructure.
On the other hand, now we no longer need to manage replica ids when
accessing nested whatevers.
In particular, as we do not expect to use more than one replica ids
outside of tests, this should be better.