give the bundle state a separate name to avoid conflicts
The bundle state is inherited by the Interface, and thus imported by
user code, thus may conflict accidentially.
This also clarifies, that the State used in the schedulers is just that,
a convenient global type definition, but not necessarily the same as
defined by the Interface – though, it has to be the same at the end.