fat jar the compiler plugin
I don’t understand how sbt handles compiler dependencies.
Actually, I also don’t understand how the compiler does it.
The command line flag allows a single jar,
and repeating that flag does not work, so what to do?

The sbt documentation on the topic states that stuff needs to be added
to the plugin configuration
https://www.scala-sbt.org/1.x/docs/Compiler-Plugins.html